requirements/production.txt
# Production requirements
-r base.txt

# Production server
gunicorn==21.2.0
uvicorn[standard]==0.23.2

# Database
psycopg2-binary==2.9.7

# Monitoring
sentry-sdk[django]==1.32.0

# Performance
django-redis==5.3.0
redis==4.6.0

# Static files
whitenoise==6.5.0

# Security
django-cors-headers==4.2.0
